requirements.txt
PyYAML==6.0
transformers==4.29.0
numpy
pandas
scikit_learn
torch
tqdm
openai>=1.0.0
backoff>=2.2.1
python-dotenv>=1.0.1
requests>=2.31.0
httpx>=0.27.0
